requirements.txt
pandas==2.2.3
openpyxl==3.1.5
pyyaml==6.0.2
opencv-python==4.11.0.86
PyAutoGUI==0.9.54
twilio==9.4.4